brew-build-bottle-pr.rb: assorted fixes
1. Check that HOMEBREW_GITHUB_API_TOKEN is set and assign GITHUB_TOKEN to it.
Otherwise, `hub` commands stall.

2. Check that HOMEBREW_GITHUB_USER is set if USER is set to 'linuxbrew'.
When assigning an assignee to the created PR, we can't use 'linuxbrew'.
It has to be the person running the command. Therefore, if HOMEBREW_GITHUB_USER is unset,
we have to make sure that we're not running in a homebrew/brew container or ask the developer
to set HOMEBREW_GITHUB_USER.

3. When specifying a HEAD branch with `hub`, we have to specify OWNER:branch.
Previously, the syntax was `remote:branch`, which does not work unless remote name is equal
to user/org name on GitHub. Here, I'm introducing a remote_user function that extracts username
for the specified remote and uses it for creating a pull request.